livekit_agents/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.1
alphashape==1.3.1
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
av==16.0.1
beautifulsoup4==4.14.3
boto3==1.42.15
botocore==1.42.15
brotli==1.2.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-log==0.4.0
colorama==0.4.6
coloredlogs==15.0.1
Crawl4AI==0.7.8
cryptography==46.0.3
cssselect==1.3.0
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
fake-http-header==0.3.5
fake-useragent==2.2.0
fastuuid==0.14.0
filelock==3.20.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-speech==2.35.0
google-cloud-texttospeech==2.33.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.3.1
litellm==1.80.5
livekit==1.0.23
livekit-agents==1.3.10
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-cartesia==1.3.10
livekit-plugins-deepgram==1.3.10
livekit-plugins-elevenlabs==1.3.10
livekit-plugins-google==1.3.10
livekit-plugins-groq==1.3.10
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.3.10
livekit-plugins-silero==1.3.10
livekit-plugins-turn-detector==1.3.10
livekit-protocol==1.1.1
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.0
onnxruntime==1.23.1
openai==2.14.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
patchright==1.56.0
pillow==12.0.0
playwright==1.57.0
pluggy==1.6.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
s3transfer==0.16.0
safetensors==0.7.0
scipy==1.16.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sounddevice==0.5.3
soupsieve==2.8.1
sympy==1.14.0
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
trimesh==4.10.1
typer==0.20.1
types-protobuf==6.32.1.20251210
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
